PUT for watch

+++ Watches/Services/WatchService.cs
@@ -70,7 +70,9 @@ namespace Watches.Services
             return watch;
         }
 
-        public async Task<bool> DeleteWatchAsync(long id)
+        public async Task<bool> UpdateWatchAsync(
+            long id, string model, string title, Gender gender, int caseSize,
+            CaseMaterial caseMaterial, long brandId, long movementId)
         {
             var watch = await _dbContext.Watches
                 .FindAsync(id);
@@ -80,6 +82,28 @@ namespace Watches.Services
                 return false;
             }
 
+            watch.Title = title;
+            watch.Model = model;
+            watch.Title = title;
+            watch.Gender = gender;
+            watch.CaseSize = caseSize;
+            watch.CaseMaterial = caseMaterial;
+            watch.BrandId = brandId;
+            watch.MovementId = movementId;
+
+            await _dbContext.SaveChangesAsync();
+            return true;
+        }
+
+        public async Task<bool> DeleteWatchAsync(long id)
+        {
+            var watch = await _dbContext.Watches
+                .FindAsync(id);
+
+            if (watch == null)
+            {
+                return false;
+            }
 
             _dbContext.Watches.Remove(watch);
             await _dbContext.SaveChangesAsync();
